pyproject.toml
[tool.poetry]
name = "bip39-monero-derive"
version = "0.1.0"
description = "Standalone utility for deriving Monero mnemonic seeds from standard BIP 39 phrases (via SLIP-0010)"
authors = ["aldum <aldum@artixlinux.org>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<=3.14"
ecdsa = "^0.19.1"

[tool.poetry.group.dev.dependencies]
stickytape = "^0.2.1"
pyinstaller = "^6.13.0"
ed25519 = "^1.5"
base58 = "^2.1.1"
pyrefly = "^0.31.0"
ruff = "^0.12.12"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pyte = "^0.8.1"
pycryptodome = "^3.17.0"
hypothesis = "^6.75.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
script = "bip39_monero_derive:main"

[tool.pytest.ini_options]
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]

[tool.ruff]
line-length = 80

[tool.pyrefly]
project-includes = ["."]
search-path = ["."]
disable-search-path-heuristics = true
